Move code to session data task
